requirements.txt
flask
gunicorn
tensorflow-cpu
scikit-learn
pandas
numpy